// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type RoleStatus = (typeof RoleStatus)[keyof typeof RoleStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
  DOCUMENT: 'DOCUMENT'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const BrandStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type BrandStatus = (typeof BrandStatus)[keyof typeof BrandStatus]


export const AttributeType = {
  DROPDOWN: 'DROPDOWN',
  RADIO: 'RADIO',
  CHECKBOX: 'CHECKBOX',
  COLOR_SWATCH: 'COLOR_SWATCH',
  IMAGE_SWATCH: 'IMAGE_SWATCH'
} as const

export type AttributeType = (typeof AttributeType)[keyof typeof AttributeType]


export const StockStatus = {
  IN_STOCK: 'IN_STOCK',
  OUT_OF_STOCK: 'OUT_OF_STOCK',
  LOW_STOCK: 'LOW_STOCK'
} as const

export type StockStatus = (typeof StockStatus)[keyof typeof StockStatus]